Remove test sub-directories.
Paths for tests were getting too long. The project is already broken up
into sub-projects. The sub-project names are already qualified by
utility or algorithm. There's no need to further sub-divide the tests
for each algorithm. It will be the same basic tests for each algorithm,
a sanity test, that will cover the waterfront, and that's about it.

Closes #20.