Unable to duplicate SHA1 test in SMHasher.
Wanted to duplicate the SHA1 test resutls from SMHasher using the
Node.js SHA1 from OpenSSL, but failed to do so.

Trying to get a feel for the buffered output from a Node.js crypto hash.
The sanity test I'm using is based on the tests in SMHasher. I'm using
it confirm my implementations against SMHasher. However, I'm not able to
confirm the results of the SMHasher SHA1. The individual results diverge
after the 20th byte has been added to the example key.

This is all a part of determining the correct endianness for the hash
output buffer. The sanity test expects little endian, but I don't know
the endianness emitted by the OpenSSL hashes. I'm using the SMHasher
SHA1 as a reference because it is legible. The OpenSSL SHA1 is written
in assembler.

I've created a test and opened an Issue at the SMHasher project.

See #28.