Create a main to test the C DJB implementation.
Created a `main` function to invoke the C DJB implementation. The
program output is the same as the `djb.c` program in the `c` directory
of the project. From this basis, I ought to be able to create a C++
binding for Node.js that I can reuse for each hash.

See #32.